third_party/blink/web_tests/platform/generic/http/tests/security/mixedContent/insecure-iframe-in-main-frame-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-frame.html' was loaded over HTTPS, but requested an insecure frame 'http://example.test:8080/security/mixedContent/resources/boring.html'. This content should also be served over HTTPS.
This test opens a window that loads an insecure iframe. We should trigger a mixed content callback because the main frame in the window is HTTPS but is running insecure content.

============== Back Forward List ==============
curr->  http://127.0.0.1:8000/security/mixedContent/insecure-iframe-in-main-frame.html
===============================================

============== Back Forward List ==============
curr->  https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-frame.html
            http://example.test:8080/security/mixedContent/resources/boring.html (in frame "<!--framePath //<!--frame0-->-->")
===============================================
